Moved table drop to its own file.
