+++ db/tables/ddl/__init__.py
@@ -1,1 +1,2 @@
 from db.tables.ddl.create import *  # noqa
+from db.tables.ddl.drop import *  # noqa
